refactor(factories): `start_route` renamed to `route`
